chore(contracts): publish the autopilot report shape to the web client
